biLerp Fix
Made a small change to the biLerp function. Previously, the yPercentage would work from the top of the grid (e.g. in a 100x100 grid, yPercentage = 0.25 would result in a Y value of 75). This has been reversed to more properly represent a grid based system and now starts from the bottom.